+++ ConnectorRhino/ConnectorRhinoWebUI/Bindings/SendBinding.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Windows.Threading;
+using System.Threading;
 using ConnectorRhinoWebUI.Utils;
 using DUI3;
 using DUI3.Bindings;
@@ -25,7 +25,7 @@ public class SendBinding : ISendBinding
   public IBridge Parent { get; set; }
   private DocumentModelStore _store;
 
-  private HashSet<string> _changedObjectIds { get; set; } = new();
+  private HashSet<string> ChangedObjectIds { get; set; } = new();
   
   public SendBinding(DocumentModelStore store)
   {
@@ -39,22 +39,22 @@ public class SendBinding : ISendBinding
     RhinoDoc.AddRhinoObject += (_, e) =>
     {
       if (!_store.IsDocumentInit) return;
-      _changedObjectIds.Add(e.ObjectId.ToString());
+      ChangedObjectIds.Add(e.ObjectId.ToString());
       RhinoIdleManager.SubscribeToIdle(RunExpirationChecks);
     };
     
     RhinoDoc.DeleteRhinoObject += (_, e) =>
     {
       if (!_store.IsDocumentInit) return;
-      _changedObjectIds.Add(e.ObjectId.ToString());
+      ChangedObjectIds.Add(e.ObjectId.ToString());
       RhinoIdleManager.SubscribeToIdle(RunExpirationChecks);
     };
     
     RhinoDoc.ReplaceRhinoObject += (_, e) =>
     {
       if (!_store.IsDocumentInit) return;
-      _changedObjectIds.Add(e.NewRhinoObject.Id.ToString());
-      _changedObjectIds.Add(e.OldRhinoObject.Id.ToString());
+      ChangedObjectIds.Add(e.NewRhinoObject.Id.ToString());
+      ChangedObjectIds.Add(e.OldRhinoObject.Id.ToString());
       RhinoIdleManager.SubscribeToIdle(RunExpirationChecks);
     }; 
   }
@@ -68,39 +68,7 @@ public class SendBinding : ISendBinding
       new RhinoLayerFilter()
     };
   }
-
-  public IDictionary<Layer, IEnumerable<object>> GroupByLayer(IEnumerable<RhinoObject> rhinoObjects, RhinoDoc doc)
-  {
-    IDictionary<Layer, IEnumerable<RhinoObject>> objsByLayer = rhinoObjects.GroupBy(o => doc.Layers.FindIndex(o.Attributes.LayerIndex)).ToDictionary(key => key.Key, value => value.AsEnumerable());
-    IDictionary<Layer, List<object>> objectByNestedLayers = new Dictionary<Layer, List<object>>();
-
-    foreach (KeyValuePair<Layer, IEnumerable<RhinoObject>> layerWithObjects in objsByLayer)
-    {
-      Layer layer = layerWithObjects.Key;
-      List<object> objects = layerWithObjects.Value as List<object>;
-
-      List<Layer> layerTree = new List<Layer>() { layer };
-
-      bool layerIsChild = layer.ParentLayerId != Guid.Empty;
-      while (layerIsChild)
-      {
-        layer = doc.Layers.FindId(layer.ParentLayerId);
-        layerIsChild = layer.ParentLayerId != Guid.Empty;
-        layerTree.Add(layer);
-      }
-
-      IDictionary<Layer, List<object>> nestedLayers = new Dictionary<Layer, List<object>>();
-      foreach (Layer node in layerTree)
-      {
-        nestedLayers = new Dictionary<Layer, List<object>>();
-        nestedLayers.Add(node, objects);
-        objects = new List<object> { node };
-      }
-    }
-
-    return objectByNestedLayers.ToDictionary(k => k.Key, v => v.Value.AsEnumerable());
-  }
-
+  
   private async void SendProgress(string modelCardId, double progress)
   {
     var args = new SenderProgress()
@@ -131,10 +99,8 @@ public class SendBinding : ISendBinding
       count++;
       convertedObjects.Add(converter.ConvertToSpeckle(rhinoObject));
       double progress = (double)count / objectsIds.Count;
-      Dispatcher.CurrentDispatcher.Invoke(() =>
-      {
-        Progress.SenderProgressToBrowser(Parent, modelCardId, progress);          
-      }, DispatcherPriority.Background);
+      Progress.SenderProgressToBrowser(Parent, modelCardId, progress);
+      Thread.Sleep(200);
     }
 
     var commitObject = new Base();
@@ -168,7 +134,7 @@ public class SendBinding : ISendBinding
   private void RunExpirationChecks()
   {
     var senders = _store.GetSenders();
-    var objectIdsList = _changedObjectIds.ToArray();
+    var objectIdsList = ChangedObjectIds.ToArray();
     var expiredSenderIds = new List<string>();
     
     foreach (var sender in senders)
@@ -180,6 +146,6 @@ public class SendBinding : ISendBinding
       }
     }
     Parent.SendToBrowser(SendBindingEvents.SendersExpired, expiredSenderIds);
-    _changedObjectIds = new HashSet<string>();
+    ChangedObjectIds = new HashSet<string>();
   }
 }
