+++ DesktopUI3/DUI3/Bindings/AccountBinding.cs
@@ -1,5 +1,4 @@
-﻿using System.Collections.Generic;
-using System.Linq;
+﻿using System.Linq;
 using JetBrains.Annotations;
 using Speckle.Core.Credentials;
 
@@ -13,6 +12,11 @@ public class AccountBinding : IBinding
   [PublicAPI]
   public Account[] GetAccounts()
   {
-    return AccountManager.GetAccounts().ToArray();
+    return AccountManager.GetAccounts().Select(
+      a =>
+      {
+        a.userInfo.avatar = null; // removing this as the get accounts call was a too large string to do "executeScriptAsync" with (this was not happening if this was a direct return from a binding call).
+        return a;
+      }).ToArray();
   }
 }
